[Frontend] Horizontal bar widgets refetch (#14448)
